+++ includes/class-wc-gateway-cardknox-googlepay.php
@@ -47,25 +47,32 @@ class WCCardknoxGooglepay extends WC_Payment_Gateway_CC
         // Load the settings.
         $this->init_settings();
 
+        $option = get_option('woocommerce_cardknox_settings');
+
         $this->enabled                          = $this->get_option('googlepay_enabled');
+        $this->google_quickcheckout             = $this->get_option('googlepay_quickcheckout');
         $this->title                            = $this->get_option('googlepay_title');
         $this->description                      = __('Pay with your Google Pay.', 'woocommerce-gateway-cardknox');
         $this->googlepay_merchant_name          = $this->get_option('googlepay_merchant_name');
         $this->googlepay_environment            = $this->get_option('googlepay_environment');
         $this->googlepay_button_style           = $this->get_option('googlepay_button_style');
-        $this->googlepay_button_type            = $this->get_option('googlepay_button_type');
         $this->capture                          = 'yes' === $this->get_option('googlepay_capture', 'yes');
         $this->authonly_status                  = $this->get_option('googlepay_auth_only_order_status');
         $this->googlepay_applicable_countries   = $this->get_option('googlepay_applicable_countries');
         $this->googlepay_specific_countries     = $this->get_option('googlepay_specific_countries');
 
         $this->wcVersion = version_compare(WC_VERSION, '3.0.0', '<');
+
         // Hooks.
         add_action('wp_enqueue_scripts', array($this, 'gpayment_scripts'));
-        add_action('woocommerce_update_options_payment_gateways_' . $this->id, array($this, 'process_admin_options'));
+        add_action('woocommerce_update_options_payment_gateways_cardknox', array($this, 'process_admin_options'));
 
         add_action('woocommerce_review_order_after_submit', array($this, 'cardknox_gpay_order_after_submit'));
         add_filter('woocommerce_available_payment_gateways', array($this, 'cardknox_allow_gpay_method_by_country'));
+
+        if (is_cart() && $this->google_quickcheckout == 'no') {
+            add_action('woocommerce_proceed_to_checkout', array($this, 'cardknox_gpay_order_after_submit'), 20);
+        }
     }
 
     /**
@@ -554,7 +561,7 @@ class WCCardknoxGooglepay extends WC_Payment_Gateway_CC
                 <div class="message message-error error gpay-error" style="display: none;"></div>
             </div>
             <div id="divGpay" class="gp hidden">
-                <iframe id="igp" class="gp" data-ifields-id="igp" data-ifields-oninit="gpRequest.initGP" src="https://cdn.cardknox.com/ifields/2.15.2302.0801/igp.htm" allowpaymentrequest sandbox="allow-popups allow-modals allow-scripts allow-same-origin
+                <iframe id="igp" class="gp" data-ifields-id="igp" data-ifields-oninit="gpRequest.initGP" src="https://cdn.cardknox.com/ifields/2.15.2405.1601/igp.htm" allowpaymentrequest sandbox="allow-popups allow-modals allow-scripts allow-same-origin
                                  allow-forms allow-popups-to-escape-sandbox allow-top-navigation" title="GPay checkout page">
                 </iframe>
             </div>