+++ includes/settings-cardknox-applepay.php
@@ -14,10 +14,6 @@ $keyDefault = 'default';
 $keyDescTip = 'desc_tip';
 $buttonPrefix = 'Apple Pay Button';
 
-global $woocommerce;
-$countries_obj   = new WC_Countries();
-$countries   = $countries_obj->__get('countries');
-
 $wc_cardknox_apple_pay_settings = array(
     $keyPrefix . '_enabled' => array(
         $keyTitle       => __('Enabled', 'woocommerce-gateway-cardknox'),
@@ -26,6 +22,16 @@ $wc_cardknox_apple_pay_settings = array(
             'yes'    => __('Yes', 'woocommerce-gateway-cardknox'),
             'no'     => __('No', 'woocommerce-gateway-cardknox'),
         ),
+    ),
+    $keyPrefix . '_quickcheckout' => array(
+        $keyTitle       => __('Disable Quick Checkout to Cart Page', 'woocommerce-gateway-cardknox'),
+        $keyType        => 'checkbox',
+        'description' => __(
+            'Disable Quick Checkout to Cart Page',
+            'woocommerce-gateway-cardknox'
+        ),
+        $keyDefault    => 'no',
+        $keyDescTip    => true,
     ),
     $keyPrefix . '_title' => array(
         $keyTitle       => __('Title', 'woocommerce-gateway-cardknox'),
@@ -66,54 +72,7 @@ $wc_cardknox_apple_pay_settings = array(
             'continue'       => __('Continue', 'woocommerce-gateway-cardknox'),
             'checkout '      => __('Checkout', 'woocommerce-gateway-cardknox'),
         ),
-    ),
-    $keyPrefix . '_capture' => array(
-        $keyTitle       => __($applePrefix . ' Capture', 'woocommerce-gateway-cardknox'),
-        'label'       => __('Capture charge immediately', 'woocommerce-gateway-cardknox'),
-        $keyType        => 'checkbox',
-        'id'          => 'apple_cardknox_capture',
-        'description' => __(
-            'If the transaction is not immediately captured for Apple Pay, it will require capturing at a later time.',
-            'woocommerce-gateway-cardknox'
-        ),
-        $keyDefault     => 'yes',
-        $keyDescTip    => true,
-    ),
-    $keyPrefix . '_auth_only_order_status' => array(
-        $keyTitle       => __('Authorize Only Order Status', 'woocommerce-gateway-cardknox'),
-        $keyType        => $selectType,
-        'class'       => 'wc-enhanced-select',
-        'id'          => 'apple_cardknox_authonly_status',
-        'description' => __(
-            'Defines the intended order status after an authorization-only transaction for Apple Pay.',
-            'woocommerce-gateway-cardknox'
-        ),
-        $keyDefault     => 'on-hold',
-        $keyDescTip    => true,
-        $keyOptions     => array(
-            'on-hold'        => __(
-                'Set order status to on-hold when payment is authorized',
-                'woocommerce-gateway-cardknox'
-            ),
-            'processing' => __(
-                'Set order status to processing when payment is authorized',
-                'woocommerce-gateway-cardknox'
-            ),
-        ),
-    ),
-    $keyPrefix . '_applicable_countries' => array(
-        $keyTitle       => __('Payment From Applicable Countries', 'woocommerce-gateway-cardknox'),
-        $keyType        => $selectType,
-        $keyOptions     => array(
-            '0'      => __('All Allowed Countries', 'woocommerce-gateway-cardknox'),
-            '1'      => __('Specific Countries', 'woocommerce-gateway-cardknox'),
-        ),
-    ),
-    $keyPrefix . '_specific_countries' => array(
-        $keyTitle       => __('Payment From Specific Countries', 'woocommerce-gateway-cardknox'),
-        $keyType        => 'multiselect',
-        $keyOptions     => $countries
-    ),
+    )
 );
 
 $GLOBALS["wc_cardknox_apple_pay_settings"] = $wc_cardknox_apple_pay_settings;