+++ changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## changelog for version Version: 1.2.68
+- Added 3D Secure integration to the checkout for added security 
+- Added the card brand logos to the Credit Card tender 
+- Added Digital Wallet (Apple Pay & google Pay) quick checkout option to cart page
+- Added tabs for Apple Pay and Google Pay on Admin Settings  
+- Fixed issues with Woo Subscription 
+- Updated tax to be send to the Gateway
+- Fixed issue with white space on Expiry
+- Updated iFields latest version 2.15.2405.1601
+- Add Tax parametrer to Gateway request 
+
 ## changelog for version 1.0.16
 
 - Add support for GooglePay